Added a route for the add question
Relates #7
