+++ src/components/pages/AddQuestion/AddQuestion.tsx
@@ -1,0 +1,15 @@
+// import React from "react";
+import { useNavigate } from "react-router-dom";
+// import Button from "../../common/Button/Button";
+// import Dropdown from "../../common/Dropdown/Dropdown";
+import AddQuestionForm from "../../common/AddQuestionForm/AddQuestionForm.tsx";
+
+function AddQuestion() {
+  return (
+    <>
+      <AddQuestionForm />
+    </>
+  );
+}
+
+export default AddQuestion;
