+++ src/routes/AppRoutes.tsx
@@ -1,8 +1,9 @@
-import React from 'react';
-import { Routes, Route } from 'react-router-dom';
+import React from "react";
+import { Routes, Route } from "react-router-dom";
 import Home from "../components/pages/Home/Home";
 import Quiz from "../components/pages/Quiz/Quiz";
 import QuestionBank from "../components/pages/QuestionBank/QuestionBank";
+import AddQuestion from "../components/pages/AddQuestion/AddQuestion";
 
 const AppRoutes: React.FC = () => {
   return (
@@ -11,8 +12,9 @@ const AppRoutes: React.FC = () => {
       <Route path="/home" element={<Home />} />
       <Route path="/quiz" element={<Quiz />} />
       <Route path="/questionbank" element={<QuestionBank />} />
+      <Route path="/questionbank/addquestion" element={<AddQuestion />} />
     </Routes>
   );
-}
+};
 
 export default AppRoutes;
